+++ application/src/ext-test/java/org/opentripplanner/ext/flex/FlexIntegrationTest.java
@@ -219,7 +219,7 @@ public class FlexIntegrationTest {
     ).buildGraph();
 
     timetableRepository.index();
-    graph.index(timetableRepository.getSiteRepository());
+    graph.index();
   }
 
   private Itinerary getItinerary(GenericLocation from, GenericLocation to, int index) {