+++ application/src/main/java/org/opentripplanner/framework/application/OTPFeature.java
@@ -20,7 +20,6 @@ public enum OTPFeature {
     false,
     "Starts a background thread to continuously publish metrics about alerts. Needs to be enabled together with `ActuactorAPI`."
   ),
-  APIBikeRental(true, false, "Enable the bike rental endpoint."),
   APIServerInfo(true, false, "Enable the server info endpoint."),
   APIUpdaterStatus(true, false, "Enable endpoint for graph updaters status."),
   IncludeEmptyRailStopsInTransfers(
@@ -135,7 +134,8 @@ public enum OTPFeature {
     true,
     "Include train Sørlandsbanen in results when searching in south of Norway. Only relevant in Norway."
   ),
-  TransferAnalyzer(false, true, "Analyze transfers during graph build.");
+  TransferAnalyzer(false, true, "Analyze transfers during graph build."),
+  TriasApi(false, true, "TRIAS API.");
 
   private static final Object TEST_LOCK = new Object();