+++ application/src/main/java/org/opentripplanner/standalone/configure/ConstructApplicationModule.java
@@ -62,6 +62,7 @@ public class ConstructApplicationModule {
     var defaultRequest = launcherRequestDecorator.intercept(routerConfig.routingRequestDefaults());
 
     var transitRoutingConfig = routerConfig.transitTuningConfig();
+    var triasApiParameters = routerConfig.triasApiParameters();
     var vectorTileConfig = routerConfig.vectorTileConfig();
     var flexParameters = routerConfig.flexParameters();
 
@@ -78,6 +79,7 @@ public class ConstructApplicationModule {
       streetLimitationParametersService,
       transitRoutingConfig,
       transitService,
+      triasApiParameters,
       vectorTileConfig,
       vehicleParkingService,
       vehicleRentalService,