+++ application/src/main/java/org/opentripplanner/transit/service/DefaultTransitService.java
@@ -39,6 +39,7 @@ import org.opentripplanner.transit.api.request.FindRoutesRequest;
 import org.opentripplanner.transit.api.request.FindStopLocationsRequest;
 import org.opentripplanner.transit.api.request.TripOnServiceDateRequest;
 import org.opentripplanner.transit.api.request.TripRequest;
+import org.opentripplanner.transit.api.request.TripTimeOnDateRequest;
 import org.opentripplanner.transit.model.basic.Notice;
 import org.opentripplanner.transit.model.basic.TransitMode;
 import org.opentripplanner.transit.model.filter.expr.Matcher;
@@ -75,6 +76,11 @@ import org.opentripplanner.utils.time.ServiceDateUtils;
  * A new instance of this class should be created for each request.
  * This ensures that the same TimetableSnapshot is used for the
  * duration of the request (which may involve several method calls).
+ * <p>
+ * There is an important exception: real-time updaters may want to query the state of unpublished
+ * timetable data from a previous real-time update. In such a case the unpublished
+ * TimetableSnapshot buffer is injected into the constructor, and the service is long-lived rather
+ * than request scoped.
  */
 public class DefaultTransitService implements TransitEditorService {
 
@@ -161,16 +167,6 @@ public class DefaultTransitService implements TransitEditorService {
   @Override
   public FeedInfo getFeedInfo(String feedId) {
     return this.timetableRepository.getFeedInfo(feedId);
-  }
-
-  @Override
-  public void addAgency(Agency agency) {
-    this.timetableRepository.addAgency(agency);
-  }
-
-  @Override
-  public void addFeedInfo(FeedInfo info) {
-    this.timetableRepository.addFeedInfo(info);
   }
 
   @Override
@@ -430,7 +426,8 @@ public class DefaultTransitService implements TransitEditorService {
       timeRange,
       numberOfDepartures,
       arrivalDeparture,
-      includeCancelledTrips
+      includeCancelledTrips,
+      TripTimeOnDate.compareByDeparture()
     );
   }
 
@@ -451,7 +448,7 @@ public class DefaultTransitService implements TransitEditorService {
   }
 
   @Override
-  public List<TripTimeOnDate> findTripTimeOnDate(
+  public List<TripTimeOnDate> findTripTimesOnDate(
     StopLocation stop,
     TripPattern pattern,
     Instant startTime,
@@ -470,6 +467,12 @@ public class DefaultTransitService implements TransitEditorService {
       arrivalDeparture,
       includeCancellations
     );
+  }
+
+  @Override
+  public List<TripTimeOnDate> findTripTimesOnDate(TripTimeOnDateRequest request) {
+    OTPRequestTimeoutException.checkForTimeout();
+    return stopTimesHelper.findTripTimesOnDate(request);
   }
 
   /**
@@ -622,16 +625,6 @@ public class DefaultTransitService implements TransitEditorService {
   @Override
   public FeedScopedId getOrCreateServiceIdForDate(LocalDate serviceDate) {
     return timetableRepository.getOrCreateServiceIdForDate(serviceDate);
-  }
-
-  @Override
-  public void addTransitMode(TransitMode mode) {
-    this.timetableRepository.addTransitMode(mode);
-  }
-
-  @Override
-  public Set<TransitMode> listTransitModes() {
-    return this.timetableRepository.getTransitModes();
   }
 
   @Override