+++ application/src/test/java/org/opentripplanner/graph_builder/module/osm/UnroutableTest.java
@@ -29,7 +29,7 @@ import org.opentripplanner.transit.model.framework.Deduplicator;
  *
  * @author abyrd
  */
-public class UnroutableTest {
+class UnroutableTest {
 
   private Graph graph;