+++ application/src/test/java/org/opentripplanner/street/integration/WalkRoutingTest.java
@@ -34,7 +34,7 @@ class WalkRoutingTest {
     roundabout = model.graph();
 
     model.timetableRepository().index();
-    roundabout.index(model.timetableRepository().getSiteRepository());
+    roundabout.index();
   }
 
   /**