+++ application/src/test/java/org/opentripplanner/updater/trip/RealtimeTestEnvironmentBuilder.java
@@ -1,9 +1,13 @@
 package org.opentripplanner.updater.trip;
 
 import static org.opentripplanner.transit.model._data.TimetableRepositoryForTest.id;
+import static org.opentripplanner.updater.trip.RealtimeTestConstants.SERVICE_DATE;
+import static org.opentripplanner.updater.trip.RealtimeTestConstants.TIME_ZONE;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.IntStream;
 import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueStore;
@@ -11,7 +15,9 @@ import org.opentripplanner.model.StopTime;
 import org.opentripplanner.model.calendar.CalendarServiceData;
 import org.opentripplanner.transit.model._data.TimetableRepositoryForTest;
 import org.opentripplanner.transit.model.framework.Deduplicator;
-import org.opentripplanner.transit.model.network.TripPattern;
+import org.opentripplanner.transit.model.framework.FeedScopedId;
+import org.opentripplanner.transit.model.site.RegularStop;
+import org.opentripplanner.transit.model.site.Station;
 import org.opentripplanner.transit.model.site.StopLocation;
 import org.opentripplanner.transit.model.timetable.Trip;
 import org.opentripplanner.transit.model.timetable.TripOnServiceDate;
@@ -19,20 +25,44 @@ import org.opentripplanner.transit.model.timetable.TripTimes;
 import org.opentripplanner.transit.model.timetable.TripTimesFactory;
 import org.opentripplanner.transit.service.TimetableRepository;
 
-public class RealtimeTestEnvironmentBuilder implements RealtimeTestConstants {
+public class RealtimeTestEnvironmentBuilder {
 
-  private final TimetableRepository timetableRepository = new TimetableRepository(
-    SITE_REPOSITORY,
-    new Deduplicator()
-  );
+  private static final FeedScopedId SERVICE_ID = TimetableRepositoryForTest.id("CAL_1");
+  private final TimetableRepositoryForTest testModel = TimetableRepositoryForTest.of();
+  private final List<RegularStop> stops = new ArrayList<>();
+  private final HashMap<String, Station> stations = new HashMap<>();
+  private final List<TripInput> tripInputs = new ArrayList<>();
+
+  RealtimeTestEnvironmentBuilder() {}
+
+  public RealtimeTestEnvironmentBuilder addTrip(TripInput... trip) {
+    for (var t : trip) {
+      addTrip(t);
+    }
+    return this;
+  }
 
   public RealtimeTestEnvironmentBuilder addTrip(TripInput trip) {
-    createTrip(trip);
-    timetableRepository.index();
+    this.tripInputs.add(trip);
     return this;
   }
 
   public RealtimeTestEnvironment build() {
+    for (var stop : stops) {
+      testModel.siteRepositoryBuilder().withRegularStop(stop);
+    }
+    for (var station : stations.values()) {
+      testModel.siteRepositoryBuilder().withStation(station);
+    }
+    var timetableRepository = new TimetableRepository(
+      testModel.siteRepositoryBuilder().build(),
+      new Deduplicator()
+    );
+
+    for (TripInput tripInput : tripInputs) {
+      createTrip(tripInput, timetableRepository);
+    }
+
     timetableRepository.initTimeZone(TIME_ZONE);
     timetableRepository.addAgency(TimetableRepositoryForTest.AGENCY);
 
@@ -41,17 +71,51 @@ public class RealtimeTestEnvironmentBuilder implements RealtimeTestConstants {
       SERVICE_ID,
       List.of(SERVICE_DATE.minusDays(1), SERVICE_DATE, SERVICE_DATE.plusDays(1))
     );
+
     timetableRepository.getServiceCodes().put(SERVICE_ID, 0);
     timetableRepository.updateCalendarServiceData(
       true,
       calendarServiceData,
       DataImportIssueStore.NOOP
     );
+    timetableRepository
+      .getAllTripPatterns()
+      .forEach(pattern -> {
+        pattern.getScheduledTimetable().setServiceCodes(timetableRepository.getServiceCodes());
+      });
 
-    return new RealtimeTestEnvironment(timetableRepository);
+    timetableRepository
+      .getAllTripPatterns()
+      .forEach(pattern -> {
+        pattern.getScheduledTimetable().setServiceCodes(timetableRepository.getServiceCodes());
+      });
+
+    timetableRepository.index();
+    return new RealtimeTestEnvironment(timetableRepository, SERVICE_DATE, TIME_ZONE);
   }
 
-  private Trip createTrip(TripInput tripInput) {
+  public RealtimeTestEnvironmentBuilder withStops(String... stopIds) {
+    this.stops.addAll(Arrays.stream(stopIds).map(id -> testModel.stop(id).build()).toList());
+    return this;
+  }
+
+  public RegularStop stop(String id) {
+    var stop = testModel.stop(id).build();
+    stops.add(stop);
+    return stop;
+  }
+
+  public RegularStop stopAtStation(String stopId, String stationId) {
+    var dflt = testModel.station(stationId).build();
+    var station = stations.getOrDefault(stationId, dflt);
+    var stop = testModel.stop(stopId).withParentStation(station).build();
+
+    stops.add(stop);
+    stations.put(stationId, station);
+    return stop;
+  }
+
+  private static Trip createTrip(TripInput tripInput, TimetableRepository timetableRepository) {
     var trip = Trip.of(id(tripInput.id()))
       .withRoute(tripInput.route())
       .withHeadsign(I18NString.of("Headsign of %s".formatted(tripInput.id())))
@@ -78,19 +142,40 @@ public class RealtimeTestEnvironmentBuilder implements RealtimeTestConstants {
 
     TripTimes tripTimes = TripTimesFactory.tripTimes(trip, stopTimes, null);
 
-    final TripPattern pattern = TimetableRepositoryForTest.tripPattern(
-      tripInput.id() + "Pattern",
-      tripInput.route()
-    )
-      .withStopPattern(
-        TimetableRepositoryForTest.stopPattern(
-          tripInput.stops().stream().map(TripInput.StopCall::stop).toList()
-        )
-      )
-      .withScheduledTimeTableBuilder(builder -> builder.addTripTimes(tripTimes))
-      .build();
+    var stopPattern = TimetableRepositoryForTest.stopPattern(
+      tripInput.stops().stream().map(TripInput.StopCall::stop).toList()
+    );
 
-    timetableRepository.addTripPattern(pattern.getId(), pattern);
+    var existingPatterns = timetableRepository
+      .getAllTripPatterns()
+      .stream()
+      .filter(p -> p.getStopPattern().equals(stopPattern))
+      .toList();
+
+    if (existingPatterns.size() > 1) {
+      throw new RuntimeException(
+        "Multiple patterns found for stop pattern %s. This indicates an error during test setup.".formatted(
+            stopPattern
+          )
+      );
+    } else if (existingPatterns.size() == 1) {
+      var pattern = existingPatterns.getFirst();
+      var newPattern = pattern
+        .copy()
+        .withScheduledTimeTableBuilder(b -> b.addTripTimes(tripTimes))
+        .build();
+      timetableRepository.addTripPattern(pattern.getId(), newPattern);
+    } else {
+      var pattern = TimetableRepositoryForTest.tripPattern(
+        tripInput.id() + "Pattern",
+        tripInput.route()
+      )
+        .withStopPattern(stopPattern)
+        .withScheduledTimeTableBuilder(builder -> builder.addTripTimes(tripTimes))
+        .build();
+
+      timetableRepository.addTripPattern(pattern.getId(), pattern);
+    }
 
     return trip;
   }